feat: persist accounts to JSON using Jackson in FileAccountRepository

+++ src/main/java/org/bankapp/utils/AccountNumberGenerator.java
@@ -1,9 +1,10 @@
 package org.bankapp.utils;
 
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 public final class AccountNumberGenerator {
-    private static final AtomicLong accountNumberGenerator = new AtomicInteger(1000000000L);
+    private static final AtomicLong accountNumberGenerator = new AtomicLong(1000000000L);
     public static String generate() {
         return "AC" + accountNumberGenerator.getAndIncrement();
     }
